+++ packages/core/src/lib/types.ts
@@ -8,6 +8,7 @@ import type { I18nConfig } from "fumadocs-core/i18n";
 import type { StructuredData } from "fumadocs-core/mdx-plugins";
 import type { ContentStorage, LoaderOptions, LoaderPluginOption, Page } from "fumadocs-core/source";
 import type { TOCItemType } from "fumadocs-core/toc";
+import type { Hono } from "hono/tiny";
 import type { MiddlewareHandler } from "hono";
 import type { ReactNode } from "react";
 import { unstable_createServerEntryAdapter } from "waku/adapter-builders";
@@ -18,6 +19,7 @@ import type {
   CreateRoot,
   CreateApi,
   CreateSlice,
+  CreateInterceptor,
 } from "waku/router/server";
 
 export type Awaitable<T> = T | Promise<T>;
@@ -84,9 +86,11 @@ export interface ServerPlugin<C extends ConfigContext = ConfigContext> {
   ) => Awaitable<PressLoaderOptions>;
 
   /** create Hono middlewares */
-  createMiddlewares?: (this: AppContext<C>) => Awaitable<MiddlewareHandler[] | undefined>;
+  createMiddlewares?: (
+    this: AppContext<C>,
+    env: { app: Hono },
+  ) => Awaitable<MiddlewareHandler[] | undefined>;
 
-  unstable_onSSGRequest?: <T>(req: Request, next: () => T) => T;
   unstable_onServerEntry?: (
     entry: ReturnType<ReturnType<typeof unstable_createServerEntryAdapter>>,
   ) => ReturnType<ReturnType<typeof unstable_createServerEntryAdapter>>;
@@ -98,6 +102,7 @@ export interface BaseRouteFns {
   createRoot: CreateRoot;
   createApi: CreateApi;
   createSlice: CreateSlice;
+  createInterceptor: CreateInterceptor;
 }
 
 export type CreatePagesResult = ReturnType<typeof createPages>;