+++ packages/core/src/plugins/llms.txt.ts
@@ -74,7 +74,7 @@ export function llmsPlugin<C extends ConfigContext = ConfigContext>(
         return res;
       });
     },
-    createMiddlewares() {
+    createMiddlewares({ app }) {
       if (this.mode === "static") return;
       const middlewares: MiddlewareHandler[] = [];
 
@@ -89,34 +89,34 @@ export function llmsPlugin<C extends ConfigContext = ConfigContext>(
       };
 
       if (autoRedirect) {
-        middlewares.push(async ({ req, redirect }, next) => {
+        middlewares.push(async ({ req }, next) => {
           if (req.path.endsWith(".md") || !isMarkdownPreferred(req.raw)) return next();
 
           const parsed = parsePathname(req.path);
           if (!parsed) return next();
           const { lang, slugs } = parsed;
 
-          const loader = await this.getLoader();
-          if (!loader.getPage(slugs, lang)) return next();
-
-          return redirect(slugsToMarkdownPath(slugs, lang).pathname);
+          const url = new URL(slugsToMarkdownPath(slugs, lang).pathname, req.url);
+          const res = await app.fetch(new Request(url));
+          if (!res.ok) return next();
+          return res;
         });
       }
 
       // API route is created under _llms.txt with force dynamic, this redirects requests like "/page.md" back to "/_llms.txt/page.md"
       if (this.mode === "dynamic") {
-        middlewares.push(async ({ req, redirect }, next) => {
+        middlewares.push(async ({ req }, next) => {
           if (!req.path.endsWith(".md")) return next();
 
           const parsed = parsePathname(req.path);
           if (!parsed || parsed.slugs[0] === "_llms.txt") return next();
           const { lang, slugs } = parsed;
 
-          const loader = await this.getLoader();
           slugs[slugs.length - 1] = slugs[slugs.length - 1]!.replace(/\.md$/, "");
-          if (!loader.getPage(slugs, lang)) return next();
-
-          return redirect(slugsToMarkdownPath(slugs, lang).pathname);
+          const url = new URL(slugsToMarkdownPath(slugs, lang).pathname, req.url);
+          const res = await app.fetch(new Request(url));
+          if (!res.ok) return next();
+          return res;
         });
       }